fix: add saml provider error
